// src/common/enums/activity.enum.ts
export enum ActivityType {
    Conference = 'conference',
    Services = 'services',
    Leisure = 'leisure',
    Causerie = 'causerie',
}

export enum TargetingCode {
    AB = 'AB',
    CD = 'CD',
    ABCD = 'ABCD',
    PreA = 'Pre A',
}

export enum ChildGroup {
    PreA = 'Pre A',
    A = 'A',
    B = 'B',
    C = 'C',
    D = 'D',
}

export enum Town {
    Douala = 'Douala',
    Edea = 'Edea',
    Yaounde = 'Yaounde',
}